+++ src/main/kotlin/snyk/common/lsp/LanguageServerWrapper.kt
@@ -1,13 +1,21 @@
 package snyk.common.lsp
 
+import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.project.ProjectManager
 import com.intellij.openapi.roots.ProjectRootManager
 import io.snyk.plugin.getCliFile
 import io.snyk.plugin.pluginSettings
+import io.snyk.plugin.snykcode.core.RunUtils
+import kotlinx.coroutines.DelicateCoroutinesApi
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.launch
 import org.eclipse.lsp4j.ClientInfo
+import org.eclipse.lsp4j.DidChangeWorkspaceFoldersParams
 import org.eclipse.lsp4j.ExecuteCommandParams
 import org.eclipse.lsp4j.InitializeParams
 import org.eclipse.lsp4j.WorkspaceFolder
+import org.eclipse.lsp4j.WorkspaceFoldersChangeEvent
 import org.eclipse.lsp4j.jsonrpc.Launcher
 import org.eclipse.lsp4j.launch.LSPLauncher
 import org.eclipse.lsp4j.services.LanguageClient
@@ -16,13 +24,16 @@ import snyk.common.EnvironmentHelper
 import snyk.common.getEndpointUrl
 import snyk.common.lsp.commands.ScanDoneEvent
 import snyk.pluginInfo
+import java.util.concurrent.TimeUnit
 
 class LanguageServerWrapper(private val lsPath: String = getCliFile().absolutePath) {
     private val gson = com.google.gson.Gson()
+    val logger = Logger.getInstance("Snyk Language Server")
 
     /**
      * The language client is used to receive messages from LS
      */
+    @Suppress("MemberVisibilityCanBePrivate")
     lateinit var languageClient: LanguageClient
 
     /**
@@ -40,30 +51,54 @@ class LanguageServerWrapper(private val lsPath: String = getCliFile().absolutePa
      */
     lateinit var process: Process
 
-    fun initialize() {
-        val snykLanguageClient = SnykLanguageClient()
-        languageClient = snykLanguageClient
-        val logLevel = if (snykLanguageClient.logger.isDebugEnabled) "debug" else "info"
-        val cmd = listOf(lsPath, "language-server", "-l", logLevel)
+    var isInitializing: Boolean = false
+    val isInitialized: Boolean
+        get() = ::languageClient.isInitialized && ::languageServer.isInitialized &&
+            ::process.isInitialized && process.info().startInstant().isPresent
 
-        val processBuilder = ProcessBuilder(cmd)
-        pluginSettings().token?.let { EnvironmentHelper.updateEnvironment(processBuilder.environment(), it) }
 
-        process = processBuilder.start()
-        launcher = LSPLauncher.createClientLauncher(languageClient, process.inputStream, process.outputStream)
-        languageServer = launcher.remoteProxy
+    @OptIn(DelicateCoroutinesApi::class)
+    internal fun initialize() {
+        try {
+            isInitializing = true
+            val snykLanguageClient = SnykLanguageClient()
+            languageClient = snykLanguageClient
+            val logLevel = if (snykLanguageClient.logger.isDebugEnabled) "debug" else "info"
+            val cmd = listOf(lsPath, "language-server", "-l", logLevel)
+
+            val processBuilder = ProcessBuilder(cmd)
+            pluginSettings().token?.let { EnvironmentHelper.updateEnvironment(processBuilder.environment(), it) }
+
+            process = processBuilder.start()
+            launcher = LSPLauncher.createClientLauncher(languageClient, process.inputStream, process.outputStream)
+            languageServer = launcher.remoteProxy
+
+            GlobalScope.launch {
+                process.errorStream.bufferedReader().forEachLine { println(it) }
+            }
+
+            launcher.startListening()
+
+            sendInitializeMessage()
+        } catch (e: Exception) {
+            logger.error(e)
+        } finally {
+            isInitializing = false
+        }
     }
 
-    fun startListening() {
-        // Start the server
-        launcher.startListening()
-    }
-
-    fun sendInitializeMessage(project: Project) {
+    private fun determineWorkspaceFolders(): List<WorkspaceFolder> {
         val workspaceFolders = mutableListOf<WorkspaceFolder>()
-        ProjectRootManager.getInstance(project)
-            .contentRoots
-            .mapNotNull { WorkspaceFolder(it.url, it.name) }.toCollection(workspaceFolders)
+        ProjectManager.getInstance().openProjects.forEach { project ->
+            ProjectRootManager.getInstance(project)
+                .contentRoots
+                .mapNotNull { WorkspaceFolder(it.url, it.name) }.toCollection(workspaceFolders)
+        }
+        return workspaceFolders.toList()
+    }
+
+    fun sendInitializeMessage() {
+        val workspaceFolders = determineWorkspaceFolders()
 
         val params = InitializeParams()
         params.processId = ProcessHandle.current().pid().toInt()
@@ -71,18 +106,44 @@ class LanguageServerWrapper(private val lsPath: String = getCliFile().absolutePa
         params.initializationOptions = getInitializationOptions()
         params.workspaceFolders = workspaceFolders
 
-        languageServer.initialize(params).get()
+        languageServer.initialize(params).get(20, TimeUnit.SECONDS)
+    }
+
+    fun updateWorkspaceFolders(project: Project, added: List<WorkspaceFolder>, removed: List<WorkspaceFolder>) {
+        RunUtils.instance.runInBackground(
+            project,
+            "Updating workspace folders for project: ${project.name}"
+        ) {
+            try {
+                ensureLanguageServerInitialized()
+                val params = DidChangeWorkspaceFoldersParams()
+                params.event = WorkspaceFoldersChangeEvent(added, removed)
+                languageServer.workspaceService.didChangeWorkspaceFolders(params)
+            } catch (e: Exception) {
+                logger.error(e)
+            }
+        }
+    }
+
+    private fun ensureLanguageServerInitialized() {
+        while (isInitializing) {
+            Thread.sleep(100)
+        }
+        if (!isInitialized) {
+            initialize()
+        }
     }
 
     fun sendReportAnalyticsCommand(scanDoneEvent: ScanDoneEvent) {
+        ensureLanguageServerInitialized()
         try {
             val eventString = gson.toJson(scanDoneEvent)
             val param = ExecuteCommandParams()
             param.command = "snyk.reportAnalytics"
             param.arguments = listOf(eventString)
             languageServer.workspaceService.executeCommand(param)
-        } catch (ignored: Exception) {
-            // do nothing to not break UX for analytics
+        } catch (e: Exception) {
+            logger.error(e)
         }
     }
 
@@ -96,6 +157,7 @@ class LanguageServerWrapper(private val lsPath: String = getCliFile().absolutePa
             endpoint = getEndpointUrl(),
             cliPath = getCliFile().absolutePath,
             token = ps.token,
+            enableTrustedFoldersFeature = "false",
             filterSeverity = SeverityFilter(
                 critical = ps.criticalSeverityEnabled,
                 high = ps.highSeverityEnabled,